MyShop/MyShop.WebUI/Controllers/BasketController.cs
using MyShop.Core.Contracts;
using MyShop.Core.Models;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;
using System.Web.Mvc;

namespace MyShop.WebUI.Controllers
{
    public class BasketController : Controller
    {
        IRepository<Customer> customerRepo;
        IBasketService basketService;
        IOrderService orderService;

        public BasketController(IBasketService basketService, IOrderService orderService, IRepository<Customer> customerRepo)
        {
            this.basketService = basketService;
            this.orderService = orderService;
            this.customerRepo = customerRepo;
        }

        // GET: Basket
        public ActionResult Index()
        {
            var basketItems = basketService.GetBasketItems(this.HttpContext);
            return View(basketItems);
        }

        public ActionResult AddToBasket(string id)
        {
            basketService.AddtoBasket(this.HttpContext, id);

            return RedirectToAction("Index");
        }

        public ActionResult RemoveFromBasket(string id)
        {
            basketService.RemoveFromBasket(this.HttpContext, id);

            return RedirectToAction("Index");
        }

        public PartialViewResult BasketSummary()
        {
            var basketSummary = basketService.GetBasketSummary(this.HttpContext);

            return PartialView(basketSummary);
        }

        [Authorize]
        public ActionResult Checkout()
        {
            var customer = customerRepo.Collection().FirstOrDefault(c => c.Email == User.Identity.Name);

            if (customer != null)
            {
                var order = new Order() {
                    FirstName = customer.FirstName,
                    LastName = customer.LastName,
                    State = customer.State,
                    Street = customer.Street,
                    City = customer.City,
                    ZipCode = customer.ZipCode,
                    Email = customer.Email
                };

                return View(order);
            }

            return RedirectToAction("Error");
        }

        [HttpPost]
        public ActionResult Checkout(Order order)
        {
            var basketItems = basketService.GetBasketItems(this.HttpContext);
            order.OrderStatus = "Creating order";
            order.Email = User.Identity.Name;

            //Processing payment

            order.OrderStatus = "Payment completed";
            orderService.CreateOrder(order, basketItems);
            basketService.ClearBasket(this.HttpContext);

            return RedirectToAction("ThankYou", new { orderId = order.Id });
        }

        public ActionResult ThankYou(string orderId)
        {
            ViewBag.OrderId = orderId;
            return View();
        }
    }
}